@@ -10,17 +10,16 @@
         }
     ],
     "require": {
-        "php": "^7.3",
+        "php": "^7.4 || 8.0",
         "friendsofphp/php-cs-fixer": "^2.16"
     },
     "require-dev": {
-        "phpstan/phpstan": "^0.12",
-        "phpstan/phpstan-phpunit": "^0.12",
-        "phpunit/phpunit": "^7.5",
-        "roave/security-advisories": "dev-master",
-        "slam/php-debug-r": "^1.6",
-        "slam/phpstan-extensions": "^4.0",
-        "thecodingmachine/phpstan-strict-rules": "^0.12"
+        "phpstan/phpstan": "^0.12.50",
+        "phpstan/phpstan-phpunit": "^0.12.16",
+        "phpunit/phpunit": "^7.5.20",
+        "slam/php-debug-r": "^1.6.1",
+        "slam/phpstan-extensions": "^5.0.2",
+        "thecodingmachine/phpstan-strict-rules": "^0.12.1"
     },
     "autoload": {
         "psr-4": {
